Don't call unittest.main() from test scripts
Rationale: the main way to run tests is ``tox -e py3x``. Having
unittest.main() in the main is redundant, but could be useful to run a
test script independently. However this breaks in tests such as the
current test_body_task.py, which has a relative import to
.mock_configuration--either we keep it like thix, tox works but not
``python test_body_task.py``, or we make it a regular import and python
works but not tox. Going for the former option.