Update unit tests
As a matter of fact, ProxQP found no solution to some of the QPs solved
in the CoM-task convergence test. I tested other solvers like DAQP, OSQP
and SCS, and OSQP is the only one that works out-of-the-box on this case
(I didn't take the corresponding setup on JVRC-1 out into a standalone
example to check that all tasks are feasible and expected to converge).